api/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
Flask==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
nodeenv==1.8.0
prisma==0.12.0
pycparser==2.21
pydantic==2.6.0
pydantic_core==2.16.1
PyJWT==2.8.0
python-dotenv==1.0.1
sniffio==1.3.0
tomlkit==0.12.3
typing_extensions==4.9.0
waitress==2.1.2
Werkzeug==3.0.1
